hydra: suppress -Wconversion
Mostly to add cast for size_t to int conversions.

+++ src/pm/hydra/mpiexec/pmiserv_pmci.c
@@ -162,7 +162,7 @@ HYD_status HYD_pmci_wait_for_completion(int timeout)
 
         while (pg_scratch->control_listen_fd != HYD_FD_CLOSED) {
             gettimeofday(&now, NULL);
-            time_elapsed = (now.tv_sec - start.tv_sec);
+            time_elapsed = (int) (now.tv_sec - start.tv_sec);
             time_left = timeout;
             if (timeout > 0) {
                 if (time_elapsed > timeout) {
@@ -191,7 +191,7 @@ HYD_status HYD_pmci_wait_for_completion(int timeout)
     /* Either all application processes exited or we have timed out.
      * We now wait for all the proxies to terminate. */
     gettimeofday(&now, NULL);
-    time_elapsed = (now.tv_sec - start.tv_sec);
+    time_elapsed = (int) (now.tv_sec - start.tv_sec);
     if (timeout > 0) {
         time_left = timeout - time_elapsed;
         if (time_left < 0)
